requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
jinja2>=3.1.5

# PDF Processing
PyMuPDF>=1.25.0
Pillow>=11.0.0

# AI
google-genai>=1.5.0

# Utilities
pydantic>=2.10.0
python-dotenv>=1.0.0

# Development
pytest>=8.3.0
httpx>=0.28.0
ruff>=0.9.0
